requirements.txt
speedtest-cli
requests
matplotlib
python-telegram-bot